movie/requirements.txt
streamlit
pandas
spacy
sentence-transformers
chromadb
dotenv
google-generativeai